requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-text-splitters
faiss-cpu
python-dotenv
tiktoken
streamlit
pypdf
rank_bm25
